requirements.txt
flask
Pillow
requests
pytest
pytest-flask
pytest-mock